+++ src/Corax/Querying/Matches/TermProviders/TermProvider.Exists.cs
@@ -197,7 +197,7 @@ namespace Corax.Querying.Matches.TermProviders
                 if ((termCount[i] & (long)TermIdMask.EnsureIsSingleMask) != 0)
                 {
                     Debug.Assert((termCount[i] & (long)TermIdMask.PostingList) != 0 || (termCount[i] & (long)TermIdMask.SmallPostingList) != 0);
-                    containersIds[i] = EntryIdEncodings.GetContainerId(termCount[i]);
+                    containersIds[i] = (long)EntryIdEncodings.GetContainerId(termCount[i]);
                     continue;
                 }
